+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fix merging RSpec DSL configuration from third-party gems. ([@pirj][])
 * Fix `RSpec/ExcessiveDocstringSpacing` false positive for multi-line indented strings. ([@G-Rath][])
 * Fix `Include` configuration for sub-departments. ([@pirj][])
+* Ignore heredocs in `RSpec/ExcessiveDocstringSpacing`. ([@G-Rath][])
 
 ## 2.5.0 (2021-09-21)